+++ tests/test_client.py
@@ -125,7 +125,7 @@ def test_launch_url_get(rp_client, launch_mode, project_name, expected_url):
 @mock.patch('reportportal_client.client.send_event')
 def test_skip_statistics(send_event, getenv):
     getenv.return_value = '1'
-    client = RPClient('http://endpoint', 'project', 'token')
+    client = RPClient('http://endpoint', 'project', 'api_key')
     client.session = mock.Mock()
     client.start_launch('Test Launch', timestamp())
     assert mock.call('start_launch', None, None) not in send_event.mock_calls
@@ -135,16 +135,16 @@ def test_skip_statistics(send_event, getenv):
 @mock.patch('reportportal_client.client.send_event')
 def test_statistics(send_event, getenv):
     getenv.return_value = ''
-    client = RPClient('http://endpoint', 'project', 'token')
+    client = RPClient('http://endpoint', 'project', 'api_key')
     client.session = mock.Mock()
     client.start_launch('Test Launch', timestamp())
     assert mock.call('start_launch', None, None) in send_event.mock_calls
 
 
 def test_clone():
-    args = ['http://endpoint', 'project', 'token']
-    kwargs = {'log_batch_size': 30, 'is_skipped_an_issue': False,
-              'verify_ssl': False, 'retries': 5,
+    args = ['http://endpoint', 'project']
+    kwargs = {'api_key': 'api_key', 'log_batch_size': 30,
+              'is_skipped_an_issue': False, 'verify_ssl': False, 'retries': 5,
               'max_pool_size': 30, 'launch_id': 'test-123',
               'http_timeout': (30, 30),
               'log_batch_payload_size': 1000000, 'mode': 'DEBUG'}
@@ -153,10 +153,10 @@ def test_clone():
     client._item_stack.append('test-322')
     cloned = client.clone()
     assert cloned is not None and client is not cloned
-    assert cloned.endpoint == args[0] and cloned.project == args[
-        1] and cloned.token == args[2]
-    assert cloned.log_batch_size == kwargs[
-        'log_batch_size'] and cloned.is_skipped_an_issue == kwargs[
+    assert cloned.endpoint == args[0] and cloned.project == args[1]
+    assert cloned.api_key == kwargs[
+        'api_key'] and cloned.log_batch_size == kwargs[
+               'log_batch_size'] and cloned.is_skipped_an_issue == kwargs[
                'is_skipped_an_issue'] and cloned.verify_ssl == kwargs[
                'verify_ssl'] and cloned.retries == kwargs[
                'retries'] and cloned.max_pool_size == kwargs[