+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Added
 - `__getstate__` and `__setstate__` methods in `RPClient` class to make it possible to pickle it, by @HardNorth
+### Changed
+- `token` field of `RPClient` class was renamed to `api_key` to maintain common convention, by @HardNorth
 
 ## [5.3.4]
 ### Added